+++ packages/wallet-adapter-core/src/AIP62StandardWallets/sdkWallets.ts
@@ -2,9 +2,16 @@ import { DevTWallet, TWallet } from "@atomrigslab/aptos-wallet-adapter";
 import { AptosConnectWallet } from "@aptos-connect/wallet-adapter-plugin";
 import { AptosStandardWallet } from "./WalletStandard";
 import { Network } from "@aptos-labs/ts-sdk";
+import { DappConfig } from "../WalletCore";
 
-export function getSDKWallets(dappConfig?: { network: Network }) {
+export function getSDKWallets(dappConfig?: DappConfig) {
   const sdkWallets: AptosStandardWallet[] = [];
+
+  // Need to check window is defined for AptosConnect
+  if (typeof window !== "undefined") {
+    sdkWallets.push(new AptosConnectWallet({ network: dappConfig?.network }));
+  }
+
   // Push production wallet if env is production, otherwise use dev wallet
   if (dappConfig?.network === Network.MAINNET) {
     // TODO twallet uses @aptos-labs/wallet-standard at version 0.0.11 while adapter uses
@@ -12,11 +19,6 @@ export function getSDKWallets(dappConfig?: { network: Network }) {
     sdkWallets.push(new TWallet() as any);
   } else {
     sdkWallets.push(new DevTWallet() as any);
-  }
-
-  // Need to check window is defined for AptosConnect
-  if (typeof window !== "undefined") {
-    sdkWallets.push(new AptosConnectWallet({ network: dappConfig?.network }));
   }
 
   return sdkWallets;