+++ packages/wallet-adapter-react/src/index.tsx
@@ -1,4 +1,6 @@
 export * from "@aptos-labs/wallet-adapter-core";
-export * from "./WalletItem";
 export * from "./WalletProvider";
+export * from "./components/AboutAptosConnect";
+export * from "./components/AptosPrivacyPolicy";
+export * from "./components/WalletItem";
 export * from "./useWallet";