+++ packages/wallet-adapter-react/CHANGELOG.md
@@ -1,5 +1,54 @@
 # @aptos-labs/wallet-adapter-react
 
+## 3.5.0
+
+### Minor Changes
+
+- 96df1f7: Added `AboutAptosConnect` headless component for building Aptos Connect education screens.
+- f23cf43: Fix adapter event communication
+
+### Patch Changes
+
+- Updated dependencies [96df1f7]
+- Updated dependencies [f23cf43]
+  - @aptos-labs/wallet-adapter-core@4.9.0
+
+## 3.4.3
+
+### Patch Changes
+
+- Updated dependencies [79a0212]
+  - @aptos-labs/wallet-adapter-core@4.8.2
+
+## 3.4.2
+
+### Patch Changes
+
+- cbbbe23: Added Dapp id to dappConfig
+- Updated dependencies [1644cfc]
+- Updated dependencies [cbbbe23]
+  - @aptos-labs/wallet-adapter-core@4.8.1
+
+## 3.4.1
+
+### Patch Changes
+
+- 0bdbb0d: Fixed the SVG attributes for the graphic of the Aptos logo.
+
+## 3.4.0
+
+### Minor Changes
+
+- 6bec234: Added `AptosPrivacyPolicy` headless component for building the privacy policy disclaimer that should be displayed below Aptos Connect login options.
+
+## 3.3.1
+
+### Patch Changes
+
+- Updated dependencies [e3df2db]
+- Updated dependencies [1580df8]
+  - @aptos-labs/wallet-adapter-core@4.8.0
+
 ## 3.3.0
 
 ### Minor Changes